+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ### May 11, 2017
 * Initial upload
   * Wizards Demo
+
+### May 15, 2017
+* Transcript Recorder
+* Style Updates
+* Header now links back to home